# docs/swagger.yaml
basePath: /api
definitions:
...
  models.DeviceAttestation:
    properties:
      attestation_data:
        description: JSON data
        type: string
      created_at:
        type: string
      device_id:
        type: string
      device_name:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      platform:
        type: string
      spiffe_id:
        description: SPIRE integration
        type: string
      trust_level:
        description: Attestation data
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: Device identification
        type: string
      verified_at:
        type: string
      workload_selector:
        type: string
    type: object
  models.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        description: Permission definition
        type: string
      resource:
        type: string
      roles:
        description: RBAC relationships
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        description: Role definition
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        description: RBAC relationships
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      device_attestations:
        description: Zero Trust fields
        items:
          $ref: '#/definitions/models.DeviceAttestation'
        type: array
      email:
        type: string
      first_name:
        description: Profile fields
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_admin:
        type: boolean
      last_name:
        type: string
      roles:
        description: RBAC relationships
        items:
          $ref: '#/definitions/models.Role'
        type: array
      sessions:
        items:
          $ref: '#/definitions/models.UserSession'
        type: array
      updated_at:
        type: string
      username:
        description: Authentication fields
        type: string
    type: object
  models.UserSession:
    properties:
      created_at:
        type: string
      device_id:
        description: Device and location tracking
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      trust_level:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_agent:
        type: string
      user_id:
        description: Session tracking
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Zero Trust Authentication MVP API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Zero Trust Auth API
  version: "1.0"
paths:
  /admin/roles:
    get:
      consumes:
      - application/json
      description: Get a list of all roles in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new role in the system
      parameters:
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created role
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Role already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - admin
  /admin/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role from the system
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated role data
        in: body
        name: role
        required: true
        schema:
          properties:
            description:
              type: string
            is_active:
              type: boolean
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated role
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"